Split secrets/waf triggering requests into new file.
I use the http.txt file as a general first go to wordlist.

I noticed over time and with some targets that it is highly likely to
trigger a WAF and get the rest of your requests blocked.

So I have moved most of the offending words into a secrets.txt file.